populate_db: Extract some functions to server_initialization.py.
